requirements.txt
pyzmq==19.0.2
protobuf==3.13.0
